Changed one of array points, added movement blocks, commented part of code, tried to make boolean
This time I mainly worked on keeping the cursor from going ontop of the
array, but it is still glitchy because it draws the fourloup after the
dot since I am not clearing. I also tried to make a boolean statement
for checking the pixels of the board, so if you had the correct pattern
you would win.